Split storage.go into multiple go files
This change makes the code easier to read and maintain.

This commit splits storage.go into:
- slab_id.go
- storage.go
- storage_health_check.go
- value_id.go
